% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R
\name{clarity_dictionary_chromote_session_close}
\alias{clarity_dictionary_chromote_session_close}
\title{Clarity Dictionary Chromote Session Close}
\usage{
clarity_dictionary_chromote_session_close(b)
}
\arguments{
\item{b}{chromote session}
}
\description{
Clarity Dictionary Chromote Session Close
}
\examples{
\dontrun{
b <- clarity_dictionary_chromote_session_open()

clarity_dictionary_chromote_session_close(b)
}
}
