% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R
\name{clarity_dictionary_chromote_session_open}
\alias{clarity_dictionary_chromote_session_open}
\title{Clarity Dictionary Chromote Session Open}
\usage{
clarity_dictionary_chromote_session_open()
}
\value{
the chromote session
}
\description{
Clarity Dictionary Chromote Session Open
}
\examples{
\dontrun{
b <- clarity_dictionary_chromote_session_open()
}
}
