% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R
\name{clarity_dictionary_database_connect}
\alias{clarity_dictionary_database_connect}
\title{Clarity Dictionary Database Connect}
\usage{
clarity_dictionary_database_connect(
  dsn = "Databricks",
  drv = odbc::odbc(),
  ...
)
}
\arguments{
\item{dsn}{name of dsn entry}

\item{drv}{driver}

\item{...}{additional arguements passed to DBI::dbConnect()}
}
\value{
a database connection
}
\description{
Clarity Dictionary Database Connect
}
\examples{
\dontrun{
con <- clarity_dictionary_database_connect(dsn = "Omicron Persi 8")
}
}
