% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R
\name{clarity_dictionary_database_disconnect}
\alias{clarity_dictionary_database_disconnect}
\title{Clarity Dictionary Database Disconnect}
\usage{
clarity_dictionary_database_disconnect(con)
}
\arguments{
\item{con}{database connection}
}
\description{
Clarity Dictionary Database Disconnect
}
\examples{
\dontrun{
con <- clarity_dictionary_database_connect(dsn = "Omicron Persi 8")

clarity_dictionary_database_disconnect(con)
}
}
