% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R
\name{clarity_dictionary_drop_all}
\alias{clarity_dictionary_drop_all}
\title{Clarity Dictionary Drop All}
\usage{
clarity_dictionary_drop_all(con)
}
\arguments{
\item{con}{database connection}
}
\description{
Clarity Dictionary Drop All
}
\examples{
\dontrun{
con <- DBI::dbConnect(
  drv = odbc::odbc(),
  dsn = "Databricks"
)

clarity_dictionary_drop_all(con)

DBI::dbDisconnect(con)
}
}
