% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R
\name{clarity_dictionary_ingest}
\alias{clarity_dictionary_ingest}
\title{Clarity Dictionary Ingest}
\usage{
clarity_dictionary_ingest(tables, b, con)
}
\arguments{
\item{tables}{table names to process}

\item{b}{chromote session}

\item{con}{database connection}
}
\value{
dataframe representing a logfile including columns table (name),
column_index (index of the column in the table), column_name, time_start
(before reading), time_inter (after read and before write),
time_end (after write), read (success or error message), write
(success or error message)
}
\description{
Clarity Dictionary Ingest
}
\examples{
\dontrun{

## 0. Setup
# - Ensure Windows OS ODBC Data Sources is configured with unexpired token
# - may have to wait for cluster startup on connect

## 1. Connect to Database
con <- clarity_dictionary_database_connect("Omicron Persi 8")

## 2. Prepare Database (if needed)
clarity_dictionary_drop_all(con)
clarity_dictionary_init(con)

## 3. Select Tables for Processing
tables_to_ingest <- clarity_dictionary_select_tables_to_ingest(con)

## 4. Setup Browser
b <- clarity_dictionary_chromote_session_open()

## 5. manual login

## 6. Ingest
clarity_dictionary_ingest(tables_to_ingest, b, con)

## 7. On Error
# Could drop the last table ingested and all associated records:
tables_to_ingest <-
  clarity_dictionary_revert_last_table(con, tables_to_ingest)

# Could also revert the entire database to a known good timestamp:
clarity_dictionary_revert_all(con, "2024-03-07T14:19:08Z")

## 8. Clean Up
clarity_dictionary_chromote_session_close(b)
clarity_dictionary_database_disconnect(con)

## 9. Query full dictionary
clarity_dictionary_select_all(con)
}
}
