% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R
\name{clarity_dictionary_init}
\alias{clarity_dictionary_init}
\title{Clarity Dictionary Initialize}
\usage{
clarity_dictionary_init(con)
}
\arguments{
\item{con}{database connection}
}
\description{
Clarity Dictionary Initialize
}
\examples{
\dontrun{
con <- DBI::dbConnect(
  drv = odbc::odbc(),
  dsn = "Databricks"
)

clarity_dictionary_init(con)

DBI::dbDisconnect(con)
}
}
