% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R
\name{clarity_dictionary_revert_all}
\alias{clarity_dictionary_revert_all}
\title{Clarity Dictionary Revert All}
\usage{
clarity_dictionary_revert_all(con, datetime)
}
\arguments{
\item{con}{database connection}

\item{datetime}{datetime for which to revert}
}
\description{
Clarity Dictionary Revert All
}
\examples{
\dontrun{
con <- DBI::dbConnect(
  drv = odbc::odbc(),
  dsn = "Databricks"
)

clarity_dictionary_revert_all(con, "2024-03-07T14:19:08Z")

DBI::dbDisconnect(con)
}
}
