% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R
\name{clarity_dictionary_revert_last_table}
\alias{clarity_dictionary_revert_last_table}
\title{Clarity Dictionary Revert Last Table}
\usage{
clarity_dictionary_revert_last_table(con, x)
}
\arguments{
\item{con}{database connection}

\item{x}{vector of table names being processed}
}
\value{
table names that still need to be processed
}
\description{
Clarity Dictionary Revert Last Table
}
\examples{
\dontrun{
clarity_table_names_remaining <-
  clarity_dictionary_revert_last_table(
    con,
    clarity_table_names_remaining
  )
}
}
