% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R
\name{clarity_dictionary_select_all}
\alias{clarity_dictionary_select_all}
\title{Clarity Dictionary Select All}
\usage{
clarity_dictionary_select_all(con, cat = FALSE)
}
\arguments{
\item{con}{database connection}

\item{cat}{cat the query instead of running it}
}
\value{
dataframe of entire clarity dictionary
}
\description{
Clarity Dictionary Select All
}
\examples{
\dontrun{
con <- DBI::dbConnect(
  drv = odbc::odbc(),
  dsn = "Databricks"
)

clarity_dictionary_select_all(con)

DBI::dbDisconnect(con)
}
}
