% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R
\name{clarity_dictionary_select_tables_to_ingest}
\alias{clarity_dictionary_select_tables_to_ingest}
\title{Clarity Dictionary Select Tables to Ingest}
\usage{
clarity_dictionary_select_tables_to_ingest(con)
}
\arguments{
\item{con}{database connection}
}
\value{
vector of table names requiring ingest
}
\description{
Clarity Dictionary Select Tables to Ingest
}
\examples{
\dontrun{
con <- DBI::dbConnect(
  drv = odbc::odbc(),
  dsn = "Databricks"
)

tables_to_ingest <-
  clarity_dictionary_select_tables_to_ingest(con)

DBI::dbDisconnect(con)
}
}
