% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{create_table}
\alias{create_table}
\title{Create Table}
\usage{
create_table(
  con,
  table = c("column", "column_ini_item_bridge", "description", "ini_item", "table",
    "type")
)
}
\arguments{
\item{con}{database connection}

\item{table}{name of table to create}
}
\value{
TRUE
}
\description{
Create Table
}
\examples{
\dontrun{
con <- DBI::dbConnect(
  drv = odbc::odbc(),
  dsn = "Databricks"
)

create_table(con, "column")

create_table(con, "column_ini_item_bridge")

create_table(con, "description")

create_table(con, "ini_item")

create_table(con, "table")

create_table(con, "type")

DBI::dbDisconnect(con)
}
}
