% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{create_table_zc_ehi_status}
\alias{create_table_zc_ehi_status}
\title{Create Table zc_ehi_status}
\usage{
create_table_zc_ehi_status(con)
}
\arguments{
\item{con}{database connection}
}
\value{
returns TRUE, invisibly
}
\description{
Create Table zc_ehi_status
}
\examples{
\dontrun{
con <- DBI::dbConnect(
  drv = odbc::odbc(),
  dsn = "Databricks"
)

create_table_zc_ehi_status(con)

DBI::dbDisconnect(con)
}
}
