% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{drop_table}
\alias{drop_table}
\title{Drop Table}
\usage{
drop_table(
  con,
  table = c("column", "column_ini_item_bridge", "description", "ini_item", "table",
    "type", "zc_ehi_status", "zc_ny")
)
}
\arguments{
\item{con}{database connection}

\item{table_name}{name of table to drop}
}
\value{
TODO
}
\description{
Drop Table
}
\examples{
\dontrun{
con <- DBI::dbConnect(
  drv = odbc::odbc(),
  dsn = "Databricks"
)

drop_table(con, "column")

drop_table(con, "column_ini_item_bridge")

drop_table(con, "description")

drop_table(con, "ini_item")

drop_table(con, "table")

drop_table(con, "type")

drop_table(con, "zc_ehi_status")

drop_table(con, "zc_ny")

DBI::dbDisconnect(con)
}
}
