% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest.R
\name{ingest_column_information_table}
\alias{ingest_column_information_table}
\title{Ingest Column Information Table}
\usage{
ingest_column_information_table(b, con, table, table_i, table_n)
}
\arguments{
\item{b}{chromote session}

\item{con}{database connection}

\item{table}{name of the table}

\item{table_i}{index of the table being processed out of table_n}

\item{table_n}{total number of tables being processed}
}
\value{
dataframe representing a logfile including columns table (name),
column_index (index of the column in the table), column_name, time_start
(before reading), time_inter (after read and before write),
time_end (after write), read (success or error message), write
(success or error message)
}
\description{
Ingest Column Information Table
}
