% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{is_wustl_login_page}
\alias{is_wustl_login_page}
\title{Verify WUSTL Login Page}
\usage{
is_wustl_login_page(b)
}
\arguments{
\item{b}{chromote session}
}
\value{
logical indicating presence of WUSTL login page
}
\description{
Verify WUSTL Login Page
}
