% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{navigate_to_table}
\alias{navigate_to_table}
\title{Navigate to Table Page}
\usage{
navigate_to_table(b, table)
}
\arguments{
\item{b}{chromote session}

\item{table}{name of table}
}
\description{
Navigate to Table Page
}
