% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest.R
\name{scrape_column_information}
\alias{scrape_column_information}
\title{Scrape Column Information}
\usage{
scrape_column_information(b, tr_i, column_i, column_n, table, table_i, table_n)
}
\arguments{
\item{b}{chromote session}

\item{tr_i}{index of the initial html row for the column}

\item{column_i}{index of the column in the clarity table}

\item{column_n}{total columns in the clarity table}

\item{table}{clarity table name}

\item{table_i}{index of the current clarity table being processed}

\item{table_n}{total number of clarity tables to be processed}
}
\value{
a list of dataframes to pass to write_column_information() for
writing to the database (table, type, ini_item, description, and column)
}
\description{
Scrape Column Information
}
