% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{select_clarity_tables}
\alias{select_clarity_tables}
\title{Select Clarity Tables}
\usage{
select_clarity_tables(con)
}
\arguments{
\item{con}{database connection}
}
\value{
dataframe of table names that may or may not have dictionary
documentation ingested but should
}
\description{
Select Clarity Tables
}
