% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{select_existing_id}
\alias{select_existing_id}
\title{Select Existing Id (Primary Key)}
\usage{
select_existing_id(con, table = c("ini_item", "table", "type"), x)
}
\arguments{
\item{con}{database connection}

\item{table}{name of table for which to select existing id}

\item{x}{one-row dataframe containing ini and item columns for table
ini_item or name column for other tables}
}
\value{
id (primary key) of the existing record or NA if not found.
}
\description{
Select Existing Id (Primary Key)
}
\examples{
\dontrun{
con <- DBI::dbConnect(
  drv = odbc::odbc(),
  dsn = "Databricks"
)

x <- tibble(id = NA_integer, ini = "EPT", item = ".1")
select_existing_id(con, "ini_item", x)

y <- tibble(id = NA_integer_, name = "PAT_ID")
select_existing_id(con, "table", y)

z <- tibble(id = NA_integer_, name = "VARCHAR(255)")
select_existing_id(con, "type", z)

DBI::dbDisconnect(con)
}
}
