% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{select_ingested_tables}
\alias{select_ingested_tables}
\title{Select Ingested Tables}
\usage{
select_ingested_tables(con)
}
\arguments{
\item{con}{database connection}
}
\value{
dataframe of table names that appear in the dictionary table table.
Note, the last table may not have been fully ingested if an error occurred.
}
\description{
Select Ingested Tables
}
