% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{select_new_id}
\alias{select_new_id}
\title{Select New Id (Primary Key)}
\usage{
select_new_id(
  con,
  table = c("column", "description", "ini_item", "table", "type")
)
}
\arguments{
\item{con}{database connection}

\item{table}{name of the table for which to select the next available id
(primary key)}
}
\value{
the next available id (primary key) for the table
}
\description{
Select New Id (Primary Key)
}
\examples{
\dontrun{
con <- DBI::dbConnect(
  drv = odbc::odbc(),
  dsn = "Databricks"
)

select_new_id(con, "column")

select_new_id(con, "description")

select_new_id(con, "ini_item")

select_new_id(con, "table")

select_new_id(con, "type")

DBI::dbDisconnect(con)
}
}
