% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{select_or_append}
\alias{select_or_append}
\title{Select or Append to a Table}
\usage{
select_or_append(con, table = c("ini_item", "table", "type"), x)
}
\arguments{
\item{con}{database connection}

\item{table}{table name for which to search for or append the record}

\item{x}{one-row dataframe representing a record to find by id or append in
the table}
}
\value{
id (primary key) of the existing or appended record
}
\description{
Select or Append to a Table
}
\examples{
\dontrun{
con <- DBI::dbConnect(
  drv = odbc::odbc(),
  dsn = "Databricks"
)

x <- tibble(id = NA_integer, ini = "EPT", item = ".1")
select_or_append(con, "ini_item", x)

y <- tibble(id = NA_integer_, name = "PAT_ID")
select_or_append(con, "table", y)

z <- tibble(id = NA_integer_, name = "VARCHAR(255)")
select_or_append(con, "type", z)

DBI::dbDisconnect(con)
}
}
