% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest.R
\name{write_column_information}
\alias{write_column_information}
\title{Write Column Information}
\usage{
write_column_information(con, data)
}
\arguments{
\item{con}{database connection}

\item{data}{list of dataframes from scrape_column_information() to write to
the database}
}
\description{
Write Column Information
}
